# Speed up integer parsing