There's no need to check for _possible_ overflow before we handle each byte of input. We can just handle the next byte and then check whether overflow occurred. This shrinks the overflow checks in half.

amd64:

```
name                             old time/op   new time/op   delta
Unmarshal/*json.codeResponse2-4   4.94ms ± 2%   4.74ms ± 1%  -4.11%  (p=0.008 n=5+5)

name                             old speed     new speed     delta
Unmarshal/*json.codeResponse2-4  393MB/s ± 2%  409MB/s ± 1%  +4.27%  (p=0.008 n=5+5)
```

arm64:

```
name                              old time/op   new time/op   delta
Unmarshal/*json.codeResponse2-10   4.43ms ± 0%   4.26ms ± 0%  -3.82%  (p=0.000 n=19+18)

name                              old speed     new speed     delta
Unmarshal/*json.codeResponse2-10  438MB/s ± 0%  455MB/s ± 0%  +3.97%  (p=0.000 n=19+18)
```